Fixed Stream guard + added translation units
